requirements.txt
numpy
opencv-python
Pillow
PyMuPDF
paddleocr
paddlepaddle=3.2.2
matplotlib
sentence-transformers
faiss-cpu
